Add patch to fix dlopen of PMIx v3 library
When slurm is built against PMIx v3, the full_path was not set,
resulting in a file not found error during dlopen(). Also, dlopen()
is hardcoded on the so file without major release number. Following
Debian packaging guidelines, this file is not installed by the PMIx
library package but its -dev variant.

To avoid a runtime dep on -dev package, try loading the shared library
with its major release number. After a quick overview a PMIx releases,
it seems the major release number of its shared library is always 2.